Move icons configuration to separate file
